Structure the program
Created the operations and the function operate that runs them given
the numers and the operator. I use different classes for different
types of buttons so that I don't need to write a bunch of if statements
in btn.addEventListener but I can write different addEventListener to
different types of buttons. Clicking the buttons will display the
content in the screen component. I need to store the numbers and
operator in variables.