Make the UI design
Completed the design of the calculator for the most part.
Still need to link my github page to the github icon and
delete the background of the "= sign".
Also need to change the html or js script to make the buttons
work when pressed. In js we are checking if the text in the buttons
are equal to the relative symbols but now there are icons there.