Complete the styling of the calculator
Now the CSS is ready, maybe only change the background/color
of the github button. The link is working.
Need to make the op-btns work when clicked with mouse.
The numbers on the numpad doesn't work (numbers on top do).
Need to add those key codes/keys in js. +, *, / work even on numapad.