FINISH THE PROJECT
Now dividing by 0 shows a message on the screen and the text
"Error" on the lastPressed div. Pressing or clicking an operator
when the result of the previous operation is a decimal number will
show a rounded result followed by the operator. The result of an
operation after pressing equal is going to be rounded as well.
Clicking and pressing should have the exact same behaviour.